#19 [feat] SharedPrefManager 인터페이스 추가 및 TodoRebootBroadcastReceiver에 AndroidEntryPoint 추가
