+++ app/src/main/java/com/gojol/notto/model/SharedPrefManager.kt
@@ -1,26 +1,8 @@
 package com.gojol.notto.model
 
-import android.content.Context
-import dagger.hilt.android.qualifiers.ApplicationContext
-import javax.inject.Inject
-import javax.inject.Singleton
+interface SharedPrefManager {
 
-@Singleton
-class SharedPrefManager @Inject constructor(@ApplicationContext context: Context) {
+    fun setBoolean(key: String, value: Boolean)
 
-    private val prefs = context.getSharedPreferences(FILE_NAME, Context.MODE_PRIVATE)
-
-    fun setBoolean(key: String, value: Boolean) {
-        prefs.edit()
-            .putBoolean(key, value)
-            .apply()
-    }
-
-    fun getBoolean(key: String): Boolean {
-        return prefs.getBoolean(key, false)
-    }
-
-    companion object {
-        private const val FILE_NAME = "pushNotification"
-    }
+    fun getBoolean(key: String): Boolean
 }
